Return failure codes from make_tests
make_tests sets result, but the code calling it expects it to set its
exit code instead ($?). This changes make_tests to return 1 instead of
setting result to 1; in addition, it will return the number of failed
tests, which allows using the test script as a whole in jobs which
expect a global failure indication.

Signed-off-by: Stephen Kitt <steve@sk2.org>