Fix python 3 compatibility in server_http.py
Building osslsigncode fails on systems with older versions of Python 3.
This is due to the server_http.py script, that runs as a part of the
test procedure. It imports the ThreadingHTTPServer-module which was
added to Python in version 3.7.

Replacing ThreadingHTTPServer with a plain HTTPServer allows us to build
for older versions of Python 3 without excluding the tests from the
build.